Moved index and info encoders to sstable
